Rename `RenameAnon` to `Rename`
In preparation for adjusting it to deal with #589.
