+++ hs-bindgen/src-internal/HsBindgen/Frontend/Pass/ResolveBindingSpec/IsPass.hs
@@ -6,7 +6,7 @@ import HsBindgen.BindingSpec qualified as BindingSpec
 import HsBindgen.Frontend.AST.Internal (CheckedMacro, ValidPass)
 import HsBindgen.Frontend.Graph.UseDecl (UseDeclGraph)
 import HsBindgen.Frontend.Pass
-import HsBindgen.Frontend.Pass.RenameAnon.IsPass
+import HsBindgen.Frontend.Pass.Rename.IsPass
 import HsBindgen.Language.C
 
 {-------------------------------------------------------------------------------
